+++ pyobs/robotic/lco/taskarchive.py
@@ -19,6 +19,7 @@ import astropy.units as u
 
 from pyobs.robotic.task import Task
 from pyobs.utils.time import Time
+from .portal import Portal
 from ..taskarchive import TaskArchive
 from .task import LcoTask
 
@@ -70,6 +71,9 @@ class LcoTaskArchive(TaskArchive):
         """
         TaskArchive.__init__(self, **kwargs)
 
+        # portal
+        self._portal = Portal(url, token)
+
         # store stuff
         self._url = url
         self._site = site
@@ -102,221 +106,7 @@ class LcoTaskArchive(TaskArchive):
 
     async def open(self) -> None:
         """Open scheduler."""
-
-        # get stuff from portal
-        await self._init_from_portal()
-
-        # start update thread
-        if self._update:
-            asyncio.create_task(self._update_schedule())
-
-    async def _portal_get(self, url: str) -> dict:
-        """Do a GET request on the portal.
-
-        Args:
-            url: URL to request.
-
-        Returns:
-            Response for request.
-
-        Raises:
-            RuntimeError if the call failed.
-            TimeoutError if the call timed out.
-        """
-        async with aiohttp.ClientSession() as session:
-            async with session.get(url, headers=self._header, timeout=10) as response:
-                if response.status != 200:
-                    raise RuntimeError("Invalid response from portal: " + await response.text())
-                return await response.json()
-
-    async def _init_from_portal(self) -> None:
-        """Initialize scheduler from portal."""
-
-        # get instruments
-        # don't catch exception, we want to fail, if something goes wrong here
-        data = await self._portal_get(urljoin(self._url, "/api/instruments/"))
-
-        # and store
-        self.instruments = {k.lower(): v for k, v in data.items()}
-
-    async def _update_schedule(self) -> None:
-        """Update thread."""
-        while True:
-            # do actual update
-            try:
-                await self._update_now()
-            except:
-                log.exception("An exception occurred.")
-
-            # sleep a little
-            await asyncio.sleep(10)
-
-    async def _update_now(self, force: bool = False) -> None:
-        """Update list of requests.
-
-        Args:
-            force: Force update.
-        """
-
-        # remember now
-        now = Time.now()
-
-        # get time of last scheduler run and check, whether we need an update, which is not the case, if
-        # - we updated before
-        # - AND last update was after last schedule update
-        # - AND last update is less then 1 min ago
-        # - AND force is set to False
-        last_scheduled = await self.last_scheduled()
-        if (
-            self._last_schedule_time is not None
-            and (last_scheduled is None or self._last_schedule_time >= last_scheduled)
-            and self._last_schedule_time > now - TimeDelta(1.0 * u.minute)
-            and force is False
-        ):
-            # need no update
-            return
-
-        # need update!
-        try:
-            tasks = await self.get_pending_tasks(
-                end_after=now, start_before=now + TimeDelta(24 * u.hour), include_running=False
-            )
-        except TimeoutError:
-            log.error("Request timed out")
-            await asyncio.sleep(60)
-            return
-        except RuntimeError:
-            log.warning("Could not fetch schedule.")
-            return
-
-        # any changes?
-        if sorted(tasks) != sorted(self._tasks):
-            log.info("Task list changed, found %d task(s) to run.", len(tasks))
-            for task_id, task in tasks.items():
-                log.info(f"  - {task.start} to {task.end}: {task.name} (#{task_id})")
-
-        # update
-        self._tasks = cast(Dict[str, LcoTask], tasks)
-
-        # finished
-        self._last_schedule_time = now
-
-    async def get_pending_tasks(
-        self, start_before: Time, end_after: Time, include_running: bool = True
-    ) -> Dict[str, Task]:
-        """Fetch pending tasks from portal.
-
-        Args:
-            start_before: Task must start before this time.
-            end_after: Task must end after this time.
-            include_running: Whether to include a currently running task.
-
-        Returns:
-            Dictionary with tasks.
-
-        Raises:
-            Timeout: If request timed out.
-            RuntimeError: If something goes wrong.
-        """
-
-        # define states
-        states = ["PENDING"]
-        if include_running:
-            states += ["IN_PROGRESS"]
-
-        # get url and params
-        url = urljoin(self._url, "/api/observations/")
-        params = {
-            "site": self._site,
-            "end_after": end_after.isot,
-            "start_before": start_before.isot,
-            "state": states,
-            "request_state": "PENDING",
-            "limit": 1000,
-        }
-
-        # do request
-        async with aiohttp.ClientSession() as session:
-            async with session.get(url, headers=self._header, params=params, timeout=10) as response:
-                if response.status != 200:
-                    raise RuntimeError("Invalid response from portal.")
-                data = await response.json()
-
-                # get schedule
-                schedules = data["results"]
-
-                # create tasks
-                tasks = {}
-                for sched in schedules:
-                    # parse start and end
-                    sched["start"] = Time(sched["start"])
-                    sched["end"] = Time(sched["end"])
-
-                    # create task
-                    task = self._create_task(LcoTask, config=sched, scripts=self.scripts)
-                    tasks[sched["request"]["id"]] = task
-
-                # finished
-                return tasks
-
-    def get_task(self, time: Time) -> Optional[LcoTask]:
-        """Returns the active task at the given time.
-
-        Args:
-            time: Time to return task for.
-
-        Returns:
-            Task at the given time or None.
-        """
-
-        # loop all tasks
-        for task in self._tasks.values():
-            # running now?
-            if task.start <= time < task.end and not task.is_finished():
-                return task
-
-        # nothing found
-        return None
-
-    async def run_task(self, task: Task) -> bool:
-        """Run a task.
-
-        Args:
-            task: Task to run
-
-        Returns:
-            Success or not
-        """
-
-        # run task
-        await task.run()
-
-        # force update tasks
-        await self._update_now(force=True)
-
-        # finish
-        return True
-
-    async def send_update(self, status_id: int, status: Dict[str, Any]) -> None:
-        """Send report to LCO portal
-
-        Args:
-            status_id: id of config status
-            status: Status dictionary
-        """
-
-        log.info("Sending configuration status update to portal...")
-        url = urljoin(self._url, "/api/configurationstatus/%d/" % status_id)
-
-        # do request
-        try:
-            async with aiohttp.ClientSession() as session:
-                async with session.patch(url, json=status, headers=self._header, timeout=10) as response:
-                    if response.status != 200:
-                        log.error("Could not update configuration status: %s", await response.text())
-
-        except TimeoutError:
-            log.error("Request timed out.")
+        await TaskArchive.open(self)
 
     async def last_changed(self) -> Optional[Time]:
         """Returns time when last time any blocks changed."""
@@ -324,31 +114,11 @@ class LcoTaskArchive(TaskArchive):
         # try to update time
         try:
             # get data
-            data = await self._portal_get(urljoin(self._url, "/api/last_changed/"))
+            self._last_changed = await self._portal.last_changed()
 
-            # get last change
-            self._last_changed = data["last_change_time"]
+        finally:
+            # even in case of errors, return last time
             return self._last_changed
-
-        except (TimeoutError, aiohttp.ClientConnectorError):
-            # in case of errors, return last time
-            return self._last_changed
-
-    async def last_scheduled(self) -> Optional[Time]:
-        """Returns time of last scheduler run."""
-
-        # try to update time
-        try:
-            # get data
-            data = await self._portal_get(urljoin(self._url, "/api/last_scheduled/"))
-
-            # get last change
-            self._last_changed = data["last_schedule_time"]
-            return self._last_changed
-
-        except (TimeoutError, aiohttp.ClientConnectorError):
-            # in case of errors, return last time
-            return self._last_scheduled
 
     async def get_schedulable_blocks(self) -> List[ObservingBlock]:
         """Returns list of schedulable blocks.
@@ -362,11 +132,11 @@ class LcoTaskArchive(TaskArchive):
             raise ValueError("No instrument type for portal set.")
 
         # get data
-        schedulable = await self._portal_get(urljoin(self._url, "/api/requestgroups/schedulable_requests/"))
+        schedulable = await self._portal.schedulable_requests()
 
         # get proposal priorities
-        data = await self._portal_get(urljoin(self._url, "/api/proposals/"))
-        tac_priorities = {p["id"]: p["tac_priority"] for p in data["results"]}
+        proposals = await self._portal.proposals()
+        tac_priorities = {p["id"]: p["tac_priority"] for p in proposals}
 
         # loop all request groups
         blocks = []
@@ -430,114 +200,6 @@ class LcoTaskArchive(TaskArchive):
 
         # return blocks
         return blocks
-
-    async def update_schedule(self, blocks: List[ObservingBlock], start_time: Time) -> None:
-        """Update the list of scheduled blocks.
-
-        Args:
-            blocks: Scheduled blocks.
-            start_time: Start time for schedule.
-        """
-
-        # create observations
-        observations = self._create_observations(blocks)
-
-        # cancel schedule
-        await self._cancel_schedule(start_time)
-
-        # send new schedule
-        await self._submit_observations(observations)
-
-    async def _cancel_schedule(self, now: Time) -> None:
-        """Cancel future schedule."""
-
-        # define parameters
-        params = {
-            "site": self._site,
-            "enclosure": self._portal_enclosure,
-            "telescope": self._portal_telescope,
-            "start": now.isot,
-            "end": (now + self._period).isot,
-        }
-
-        # url and headers
-        url = urljoin(self._url, "/api/observations/cancel/")
-        headers = {"Authorization": "Token " + self._token, "Content-Type": "application/json; charset=utf8"}
-
-        # cancel schedule
-        log.info("Deleting all scheduled tasks after %s...", now.isot)
-        async with aiohttp.ClientSession() as session:
-            async with session.post(url, json=params, headers=headers, timeout=10) as response:
-                if response.status != 200:
-                    log.error("Could not cancel schedule: %s", await response.text())
-
-    def _create_observations(self, blocks: List[ObservingBlock]) -> List[Dict[str, Any]]:
-        """Create observations from schedule.
-
-        Args:
-            blocks: List of scheduled blocks
-
-        Returns:
-            List with observations.
-        """
-
-        # loop blocks
-        observations = []
-        for block in blocks:
-            # get request
-            request = block.configuration["request"]
-
-            # add observation
-            observations.append(
-                {
-                    "site": self._site,
-                    "enclosure": self._portal_enclosure,
-                    "telescope": self._portal_telescope,
-                    "start": block.start_time.isot,
-                    "end": block.end_time.isot,
-                    "request": request["id"],
-                    "configuration_statuses": [
-                        {
-                            "configuration": request["configurations"][0]["id"],
-                            "instrument_name": self._portal_instrument,
-                            "guide_camera_name": self._portal_instrument,
-                        }
-                    ],
-                }
-            )
-
-        # return list
-        return observations
-
-    async def _submit_observations(self, observations: List[Dict[str, Any]]) -> None:
-        """Submit observations.
-
-        Args:
-            observations: List of observations to submit.
-        """
-
-        # nothing?
-        if len(observations) == 0:
-            return
-
-        # url and headers
-        url = urljoin(self._url, "/api/observations/")
-        headers = {"Authorization": "Token " + self._token, "Content-Type": "application/json; charset=utf8"}
-
-        # submit obervations
-        async with aiohttp.ClientSession() as session:
-            async with session.post(url, json=observations, headers=headers, timeout=10) as response:
-                if response.status != 200:
-                    log.error("Could not submit observations: %s", await response.text())
-                data = await response.json()
-
-        # log
-        log.info("%d observations created.", data["num_created"])
-
-        # errors?
-        if "errors" in data and len(data["errors"] > 0):
-            for err in data["errors"].values():
-                log.warning("Error from portal: %s", err["non_field_errors"])
 
 
 __all__ = ["LcoTaskArchive"]
