+++ pyobs/robotic/taskarchive.py
@@ -4,7 +4,7 @@ from astroplan import ObservingBlock
 
 from pyobs.utils.time import Time
 from .task import Task
-from ..object import Object
+from pyobs.object import Object
 
 
 class TaskArchive(Object, metaclass=ABCMeta):
@@ -26,70 +26,11 @@ class TaskArchive(Object, metaclass=ABCMeta):
         ...
 
     @abstractmethod
-    async def last_scheduled(self) -> Optional[Time]:
-        """Returns time of last scheduler run."""
-        ...
-
-    @abstractmethod
     async def get_schedulable_blocks(self) -> List[ObservingBlock]:
         """Returns list of schedulable blocks.
 
         Returns:
             List of schedulable blocks
-        """
-        ...
-
-    @abstractmethod
-    async def update_schedule(self, blocks: List[ObservingBlock], start_time: Time) -> None:
-        """Update the list of scheduled blocks.
-
-        Args:
-            blocks: Scheduled blocks.
-            start_time: Start time for schedule.
-        """
-        ...
-
-    @abstractmethod
-    async def get_pending_tasks(
-        self, start_before: Time, end_after: Time, include_running: bool = True
-    ) -> Dict[str, Task]:
-        """Fetch pending tasks from portal.
-
-        Args:
-            start_before: Task must start before this time.
-            end_after: Task must end after this time.
-            include_running: Whether to include a currently running task.
-
-        Returns:
-            Dictionary with tasks.
-
-        Raises:
-            Timeout: If request timed out.
-            ValueError: If something goes wrong.
-        """
-        ...
-
-    @abstractmethod
-    def get_task(self, time: Time) -> Optional[Task]:
-        """Returns the active task at the given time.
-
-        Args:
-            time: Time to return task for.
-
-        Returns:
-            Task at the given time.
-        """
-        ...
-
-    @abstractmethod
-    async def run_task(self, task: Task) -> bool:
-        """Run a task.
-
-        Args:
-            task: Task to run
-
-        Returns:
-            Success or not
         """
         ...
